+++ src/core/as.js
@@ -237,10 +237,11 @@ async function processDataForRPInternalAsync(
 }
 
 export async function afterGotDataFromCallback(
-  { response, body },
+  { error, response, body },
   additionalData
 ) {
   try {
+    if (error) throw error;
     if (response.status === 204) {
       await db.setRPIdFromRequestId(
         additionalData.requestId,